clientlib: Moved client_handler and modified API.
[GH-25]

All the methods that rely on the access to the DB are private now.
Eventually, the API will decide whether to call those or provide data
from the cache.
